refactor: move operation handlers to a separate sub-package
